+++ modules/main.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
-import timeit
 import hashlib
 import hmac
 import os
-import json
 import getpass
 import sys
 
@@ -19,57 +17,58 @@ class Manager:
             self.run()
 
         else:
-            if os.path.isfile('db/info.json'):
-                with open("db/info.json", 'r') as f:
-                    jfile = json.load(f)
+            with sqlite3.connect('vault.db') as db: 
+                cursor = db.cursor()
 
-                self.master_pw = getpass.getpass(
-                    'Enter your master password: ')
+            if os.path.isfile('vault.db'):
+                try: 
+                    self.master_pw = getpass.getpass('Enter your master password: ')
 
-                h = hmac.new(self.master_pw.encode(), msg=str(
-                    jfile["Informations"]["salt"]).encode(), digestmod=hashlib.sha3_512).hexdigest()
+                    cursor.execute("SELECT * FROM masterpassword")
+                    for row in cursor.fetchall():
+                        master = row[0]
+                        salt = row[1] 
 
-                if h == jfile["Informations"]["master_password"]:
-                    Menu(self.master_pw).menu_interface()
-                    self.run()
+                    h = hmac.new(self.master_pw.encode(), msg=str(salt).encode(), digestmod=hashlib.sha3_512).hexdigest()
+                    if h == master:
+                        Menu(self.master_pw).menu_interface()
+                        self.run()
+                    else:
+                        print(Fore.RED + 'The master password is not correct.\033[0m')
+                        self.main()
+                    
+                except sqlite3.Error:
+                    print(Fore.GREEN + 'To start, you have to create a master password. Be careful not to lose it as it is unrecoverable.' + Style.RESET_ALL)
+                    master_pw = getpass.getpass('Create a master password to use the program: ')
+                    verify_master = getpass.getpass('Enter your master password again to verify it: ')
+                    time.sleep(1)
+                    
+                    if master_pw == verify_master:
+                        if master_pw.isnumeric() or master_pw.isspace():
+                            print(Fore.RED + '\nThe password is not correct. Please try again.' + Style.RESET_ALL)
+                            return self.main()
 
-                else:
-                    print(Fore.RED + 'The master password is not correct.\033[0m')
-                    self.main()
+                        elif len(master_pw) < 8:
+                            print(Fore.RED + '\nThe password must have at least 8 caracters.' + Style.RESET_ALL)
+                            return self.main()
 
-            else:
-                print(Fore.GREEN + 'To start, you have to create a master password. Be careful not to lose it as it is unrecoverable.' + Style.RESET_ALL)
-                master_pw = getpass.getpass('Create a master password to use the program: ')
-                verify_master = getpass.getpass('Enter your master password again to verify it: ')
-                time.sleep(1)
-                
-                if master_pw == verify_master:
-                    if master_pw.isnumeric() or master_pw.isspace():
-                        print(Fore.RED + '\nThe password is not correct. Please try again.' + Style.RESET_ALL)
-                        return self.main()
+                        else:
+                            with sqlite3.connect('vault.db') as db: 
+                                cursor = db.cursor()
+                                cursor.execute("CREATE TABLE masterpassword (password TEXT NOT NULL, salt TEXT NOT NULL);")
 
-                    elif len(master_pw) < 8:
-                        print(Fore.RED + '\nThe password must have at least 8 caracters.' + Style.RESET_ALL)
-                        return self.main()
+                            salt = "".join(random.choice(string.ascii_uppercase + string.digits + string.ascii_lowercase) for _ in range(32))
+                            master = hmac.new(master_pw.encode(), msg=str(salt).encode(), digestmod=hashlib.sha3_512).hexdigest()
+                            
+                            cursor.execute(f"INSERT INTO masterpassword VALUES('{master}', '{salt}')")
+                            db.commit()
+                            
+                            sys.exit(
+                                Fore.GREEN + "\nThank you! Restart the program and enter your master password to begin." + Style.RESET_ALL)
 
                     else:
-                        try:
-                            os.mkdir('db/')
-                        except Exception:
-                            pass
-
-                        jfile = {"Informations": {}}
-                        jfile["Informations"]["salt"] = "".join(random.choice(string.ascii_uppercase + string.digits + string.ascii_lowercase) for _ in range(32))
-                        jfile["Informations"]["master_password"] = hmac.new(master_pw.encode(), msg=str(jfile["Informations"]["salt"]).encode(), digestmod=hashlib.sha3_512).hexdigest()
-                        with open('db/info.json', 'w', encoding='utf-8') as jsondata:
-                            json.dump(jfile, jsondata, sort_keys=True, indent=4)
-
-                        sys.exit(
-                            Fore.GREEN + "\nThank you! Restart the program and enter your master password to begin." + Style.RESET_ALL)
-
-                else:
-                    print(Fore.RED + '\nPassword do not match. Please try again.' + Style.RESET_ALL)
-                    return self.main()
+                        print(Fore.RED + '\nPassword do not match. Please try again.' + Style.RESET_ALL)
+                        return self.main()
 
     def run(self):
         """Declarate main() and runtime
@@ -81,13 +80,9 @@ class Manager:
         def runtime():
             time.sleep(2)
             print()
-            start = timeit.default_timer()
             print("*" * 41)
             print(Fore.CYAN + 'PassVault' + Style.RESET_ALL + ' finished.')
-
-            end = timeit.default_timer()
-            print("\n[+] RUNTIME: %f seconds\n" % (end - start))
-            print("[+] Developed by vlHan. Thanks for using!")
+            print("Developed by vlHan. Thanks for using!")
             print("*" * 41)
 
         if options == 'y':