+++ CHANGELOG.md
@@ -6,8 +6,8 @@ All notable change to this project will be documented here.
 
 ## [V1.0] - 2021-12-29
 ### Added
-- db folder in .gitgnore
-- install of dependencies in run.py
+- db folder in `.gitgnore`
+- install of dependencies in `run.py`
 - variables which is intended to be private
 
 ### Changed
@@ -17,3 +17,16 @@ All notable change to this project will be documented here.
 - Imports from database.py 
 - email verification in begin_informations method
 - sqlite parameter (close database)
+
+## [V1.1] - 2021-12-30
+### Added
+- master password table (`masterpassword`)
+- a logo (`.github/logo.png`) and moved `passvault.png` to `.github` folder 
+
+### Changed
+- The json file as a file database to a table in SQlite.
+- key system to a ID sytem. 
+
+### Removed
+- json file imports, and function to stored the master password.
+- key system